Backend/requirements.txt
google-adk
openai
mysql-connector-python
python-dotenv
install 
schedule 
psutil 
pywin32